+++ project/stm32f407_bb_csr8x11_bt/bt_phybusif_h4.c
@@ -74,7 +74,7 @@ uint8_t hw_uart_bt_init(uint32_t baud_rate,uint8_t reconfig)
  * return      : NULL
  * description : The send function of USART2
 ******************************************************************************/
-void uart_bt_send(uint8_t *buf,uint8_t len)
+void uart_bt_send(uint8_t *buf,uint16_t len)
 {
     uint8_t index;
     for(index = 0; index < len ; index++)
